+++ python/egglog/builtins.py
@@ -5,7 +5,7 @@ Builtin sorts and function to egg.
 
 from __future__ import annotations
 
-from functools import partial
+from functools import partial, reduce
 from types import FunctionType
 from typing import TYPE_CHECKING, Generic, Protocol, TypeAlias, TypeVar, Union, cast, overload
 
@@ -25,13 +25,16 @@ __all__ = [
     "Bool",
     "BoolLike",
     "Map",
+    "MapLike",
     "PyObject",
     "Rational",
     "Set",
+    "SetLike",
     "String",
     "StringLike",
     "UnstableFn",
     "Vec",
+    "VecLike",
     "f64",
     "f64Like",
     "i64",
@@ -285,6 +288,22 @@ class Map(Expr, Generic[T, V], builtin=True):
     def rebuild(self) -> Map[T, V]: ...
 
 
+TO = TypeVar("TO")
+VO = TypeVar("VO")
+
+converter(
+    dict,
+    Map,
+    lambda t: reduce(
+        (lambda acc, kv: acc.insert(convert(kv[0], get_type_args()[0]), convert(kv[1], get_type_args()[1]))),
+        t.items(),
+        Map[get_type_args()].empty(),  # type: ignore[misc]
+    ),
+)
+
+MapLike: TypeAlias = Map[T, V] | dict[TO, VO]
+
+
 class Set(Expr, Generic[T], builtin=True):
     @method(egg_fn="set-of")
     def __init__(self, *args: T) -> None: ...
@@ -316,6 +335,17 @@ class Set(Expr, Generic[T], builtin=True):
 
     @method(egg_fn="rebuild")
     def rebuild(self) -> Set[T]: ...
+
+
+converter(
+    set,
+    Set,
+    lambda t: Set[get_type_args()[0]](  # type: ignore[misc,operator]
+        *(convert(x, get_type_args()[0]) for x in t)
+    ),
+)
+
+SetLike: TypeAlias = Set[T] | set[TO]
 
 
 class Rational(Expr, builtin=True):
@@ -418,7 +448,16 @@ class Vec(Expr, Generic[T], builtin=True):
     def set(self, index: i64Like, value: T) -> Vec[T]: ...
 
 
-converter(tuple, Vec, lambda t: Vec(*(convert(x, get_type_args()[0]) for x in t)))
+for sequence_type in (list, tuple):
+    converter(
+        sequence_type,
+        Vec,
+        lambda t: Vec[get_type_args()[0]](  # type: ignore[misc,operator]
+            *(convert(x, get_type_args()[0]) for x in t)
+        ),
+    )
+
+VecLike: TypeAlias = Vec[T] | tuple[TO, ...] | list[TO]
 
 
 class PyObject(Expr, builtin=True):