+++ docs/changelog.md
@@ -12,7 +12,7 @@ _This project uses semantic versioning_
 - Add working loopnest test and rewrite array api suport to be more general
 - Improve tracebacks on failing conversions.
 - Use `add_note` for exception to add more context, instead of raising a new exception, to make it easier to debug.
-- Add conversions from vecs (TODO)
+- Add conversions from generic types to be supported at runtime and typing level (so can go from `(1, 2, 3)` to `TupleInt`)
 - Open files with webbrowser instead of internal graphviz util for better support
 
 ## 8.0.1 (2024-10-24)